+++ name_id_mapping_sheet.json
@@ -1,0 +1,86 @@
+{
+    "GD0001": "Giovanni Dessert",
+    "MP0002": "Maureen Petica",
+    "JZ0003": "Jennifer Zhumi",
+    "IC0004": "Ivan Cervantes",
+    "SM0005": "Sytricia Mears",
+    "AS0006": "Andrea Serrano",
+    "JM0007": "Jennilee Mirtil",
+    "RK0008": "Rose Kalu Igwe",
+    "AP0009": "Amber Polakovich",
+    "MM0010": "Mariana Molero",
+    "JV0011": "Jessica Vargas",
+    "AZ0012": "Amanda Zamot",
+    "MC0013": "Mikaila Corriders",
+    "AG0014": "Alitzel Gonzalez",
+    "SG0015": "Shardae Gilchrist",
+    "ST0016": "Safina Theard-Lewis",
+    "KL0017": "Kaylee Lamarche",
+    "YH0018": "Yamely Hernandez Diaz",
+    "RE0019": "Raissa Ebeh",
+    "AG0020": "Alexis Geneve",
+    "NM0021": "Najara Mizhquiri",
+    "BD0022": "Barbara Djan",
+    "NZ0023": "Natalie Zuniga",
+    "GG0024": "Gianny Guzman",
+    "NM0025": "Natasha Mohan",
+    "MG0026": "Mohamed Ghanem",
+    "DC0027": "Diana Cortez",
+    "MY0028": "Mykhal Young",
+    "BB0029": "Barbara Boirond",
+    "R0030": "Rashail",
+    "SC0031": "Stephanie Cano",
+    "AJ0032": "Adama Jaiteh",
+    "KR0033": "Kylana Rayne Laspina",
+    "Jk0034": "Jianee kiary carrasco",
+    "FG0035": "Fatima Galal El-Din",
+    "AD0036": "Alison De Leon",
+    "JA0037": "José Andrés Suarez Tejada",
+    "JE0038": "Joshua Enobiomano",
+    "D0039": "DougenyFrancois",
+    "B0040": "Briana",
+    "MO0041": "Mariam Ouattara",
+    "ND0042": "Naa Dei",
+    "LG0043": "Le Garland Parker",
+    "MR0044": "MiElena Rodriguez",
+    "MM0045": "Mareda Michael",
+    "AC0046": "Angelisa Cortorreal",
+    "AR0047": "Ashley Rossis",
+    "AT0048": "Anjolaoluwa Tega",
+    "ND0049": "Naa Dei Ashie",
+    "YM0050": "Yandel Morel",
+    "HM0051": "Hector Medrano",
+    "KD0052": "Keziah Diego",
+    "OS0053": "Oumou Sylla",
+    "PG0054": "Pamela Gonzalez",
+    "GD0055": "Giovanni Dessert",
+    "RN0056": "Robert Novo",
+    "L0057": "Le'Garland",
+    "Al0058": "Alexus lawrence",
+    "JS0059": "Jianina Suazo",
+    "IR0060": "Izadeli Rivera",
+    "BO0061": "Britney Omene",
+    "EB0062": "Ernesto Barreto",
+    "JC0063": "Jennifer Calderon",
+    "CR0064": "Christiana Raphael",
+    "GR0065": "Genesis Rodriguez",
+    "L0066": "Lucia",
+    "BA0067": "Bahey Abou-Hussein",
+    "RP0068": "Reham Perry",
+    "DD0069": "Dion Dixie Oubrey Go",
+    "E0070": "Ena",
+    "BT0071": "Briana Turner",
+    "AL0072": "Alexandra Lenz",
+    "VG0073": "Vanessa Gonzalez",
+    "DM0074": "Destinie McDonald",
+    "FS0075": "Faith Singh",
+    "CB0076": "Christina Brown",
+    "JG0077": "Johanna Gomez",
+    "SI0078": "Sarah Israel",
+    "JL0079": "Joshua Levy",
+    "S0080": "Seck",
+    "SH0081": "Sarah Hoyos",
+    "SF0082": "Samuel Fakanle",
+    "JT0083": "Jessica Tapia",
+    "DZ0084": "Dayanna Zuluaga"
+}
