+++ properties.json
@@ -2,8 +2,17 @@
   "survey_monkey_spreadsheet_range": "J1:EN1000",
   "score_sheet_response_range": "Sheet1!B1:EN1000",
   "score_sheet_objective_range": "B259:AU368",
+  "score_sheet_objective_ranges": {
+    "Objective 1": "B259:B286",
+    "Objective 2": "B287:B314",
+    "Objective 3": "B315:B328",
+    "Objective 4": "B329:B342",
+    "Objective 5": "B343:B354",
+    "Objective 6": "B355:B368"
+  },
   "mentee_objective_score_sheet_range": "Cohort!D12:Q53",
   "csv_file_path": "resources/test/SurveyMonkey Data - Jeremy Test.csv",
+  "score_sheet_template_name": "Score Sheet Template",
   "score_sheet_template_id": "1EBmFujKhO84_DEZye3-kPag_-G2pWU5pUec_8YbAp3Q",
   "mentee_objective_scoresheet_id": "1GbX4ja3mPaAzbqI3STzzSWwCbdX-WOYf33i9y2F2zzA"
 }
